Fix bug where update() could post stale metadata
This is because before updating the file, we got
the pre-update metadata with a GET, and used this
to package the actual file contents. If the metadata
changed (e.g. parents change from a mv or something),
we would overwrite this when the upload completes with
the stale metadata.

The solution is to avoid packaging metadata in the
request by not using GET.