Early attempt at decoupling interdependencies of File class
Toward this end, we move most of the functionality which
deals with file fragments / file content in general into
a separate "FileContent" class. We also offload some of the
'playOnMetadata' functionality to the SimpleFileMetadata class.

Note: this commit changes the semantics of how we determine
whether the cached metadata has been initialized. Previously,
we did so by checking whether the whole metadata object was
null. Now, we prohibit null metadata values, and instead check
to see whether an Optional<Long> asof timestamp contains a
value. If it doesn't, the metadata hasn't been initialized.